% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_matching.R
\name{survey_matching}
\alias{survey_matching}
\title{Matching SoftWare for Causal Inference with Complex Survey Design}
\usage{
survey_matching(
  treatment,
  covariates,
  data,
  design,
  ps.method = c("M", "SW"),
  distance.threshold,
  M...
)
}
\arguments{
\item{treatment}{a binary \code{variable} where 1 indicates group of treatment and 0 indicates group of control}

\item{covariates}{a \code{list} of pre-treatment covariates}

\item{data}{a \code{data.frame}}

\item{design}{a \code{svydesign} object created by the svydesign function using survey package}

\item{ps.method}{This argument specifies a matching method. Currently, 'M'(nearest neighbor matching, the weight adjustment will do automatically), 'SW'(stablized weighting)}

\item{distance.threshold}{The threshold value for matched case selection. If you choose 'M', then you have to give a value for this argument.}

\item{M}{The maximum number of control cases}
}
\value{
\describe{
\item{\code{balanced_data}}{a \code{data.frame} of data balanced with propensity score methods.If "M" is used, then the weights variable is either
original weight variable or "MA_weight" for weight adjustment method. If 'SW' is used, then "SW_weight", "IPTW_weight" is the weight for un-weighted data,
using stablized weight method and inverse probability weighting method, and "SW_longwt", "IPTW_longwt" is the corresponding weight for complex survey data.  }
}
}
\description{
Matching SoftWare for Causal Inference with Complex Survey Design
}
\details{
This function estimates the RR using parametric models with complex survey design allowed
}
\examples{
library(survey)
library(surveyRR)
data("mepsRR")
options(survey.lonely.psu="adjust")
mycovariates<-c( "AGELAST"   ,   "race"   ,      "sex"   ,  "diabetes"  ,   "asthma"    ,   "heartdisease", "health"    ,
"mnhlth"   ,   "BMILAST"    ,  "education"  ,  "employed"    , "insurance"  )
CTdsgn <- svydesign(
 id = ~VARPSU,
 strata = ~VARSTR,
 weights = ~LONGWT,
 data = mepsRR,
 nest = TRUE)
 survey_matching(treatment="binary_flushot",covariates=mycovariates,
 data=mepsRR,design=CTdsgn,ps.method="M",distance.threshold=10^-5,M=2)

}
\references{
Assessing the Effectiveness of Influenza Vaccination – influence function approach for comparative studies using publicly available medical expenditure data.
Mingmei Tian, Jihnhee Yu

Xu S, Ross C, Raebel MA, et al. Use of stabilized inverse propensity scores as weights to directly estimate relative risk and its confidence intervals. Value in Health 2010; 13: 273-277.

Reardon SF, Cheadle JE and Robinson JP. The effect of Catholic schooling on math and reading development in kindergarten through fifth grade. Journal of Research on Educational Effectiveness 2009; 2: 45-87.
}
\author{
Mingmei Tian
}
