% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IF_model.R
\name{svyglm_rr_if}
\alias{svyglm_rr_if}
\title{Relative Risk and Disease Rate for Parametric model}
\usage{
svyglm_rr_if(
  treatment,
  response,
  data,
  design,
  formula,
  model = c("logit", "log-binominal", "probit"),
  ...
)
}
\arguments{
\item{treatment}{a binary \code{variable} where 1 indicates group of treatment and 0 indicates group of control}

\item{response}{a binary \code{variable} where 1 indicates group of event and 0 indicates group of non-event}

\item{data}{a \code{data.frame}}

\item{design}{a \code{svydesign} object created by the svydesign function using survey package}

\item{formula}{a model formula}

\item{model}{: a description of distribution}
}
\value{
\describe{
\item{\code{result}}{a \code{data.frame} of estimated RR and treatment rate}
}
}
\description{
Estimate relative risk using parametric models with complex survey design
}
\details{
This function estimates the RR using parametric models with complex survey design allowed
}
\examples{
library(survey)
library(surveyRR)
data("mepsRR")
options(survey.lonely.psu="adjust")
CTdsgn <- svydesign(
 id = ~VARPSU,
 strata = ~VARSTR,
 weights = ~LONGWT,
 data = mepsRR,
 nest = TRUE)
 logit_rr_if(treatment="binary_flushot",response="disease_status",data=mepsRR,design=CTdsgn,
             formula=disease_status~binary_flushot)

}
\references{
Assessing the Effectiveness of Influenza Vaccination – influence function approach for comparative studies using publicly available medical expenditure data.
Mingmei Tian, Jihnhee Yu
}
\author{
Mingmei Tian
}
